+++ src/main/java/be/baur/sds/parsing/Attribute.java
@@ -1,4 +1,4 @@
-package be.baur.sds.serialization;
+package be.baur.sds.parsing;
 
 /**
  * Attributes allowed by the SDS syntax. The lower-case name of an attribute