+++ src/test/java/test/validation/Choices.java
@@ -1,8 +1,5 @@
 package test.validation;
 
-import java.io.File;
-import java.io.InputStream;
-import java.io.InputStreamReader;
 import java.util.Iterator;
 
 import be.baur.sda.DataNode;
@@ -24,11 +21,9 @@ public final class Choices {
 		});
 		
 
-		DataNode doc = SDA.parse(new File(Choices.class.getResource("/mgtest.sda").getFile()));
-
-		InputStream sds = Choices.class.getResourceAsStream("/choices.sds");
-		Validator validator = SDS.parse(new InputStreamReader(sds, "UTF-8")).newValidator();
-
+		DataNode doc = SDA.parse(Test.getResourceFile("/mgtest.sda"));
+		Validator validator = SDS.parse(Test.getResourceFile("/choices.sds")).newValidator();
+		
 		Errors errors = validator.validate(doc);
 		//for (Error error : errors) System.out.println(error.toString());
 		Iterator<?> e = errors.iterator();