+++ src/test/java/test/validation/Contacts.java
@@ -1,8 +1,5 @@
 package test.validation;
 
-import java.io.File;
-import java.io.InputStream;
-import java.io.InputStreamReader;
 import java.util.Iterator;
 
 import be.baur.sda.DataNode;
@@ -24,10 +21,9 @@ public final class Contacts {
 		});
 		
 
-		DataNode doc = SDA.parse(new File(Contacts.class.getResource("/contacts.sda").getFile()));
-
-		InputStream sds = Contacts.class.getResourceAsStream("/contacts.sds");
-		Validator validator = SDS.parse(new InputStreamReader(sds, "UTF-8")).newValidator();
+		DataNode doc = SDA.parse(Test.getResourceFile("/contacts.sda"));
+		Validator validator = SDS.parse(Test.getResourceFile("/contacts.sds")).newValidator();
+		
 		Errors errors = validator.validate(doc);
 		//for (Error error : errors) System.out.println(error.toString());
 		Iterator<?> e = errors.iterator();